backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==2.1.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
gunicorn==21.2.0
idna==3.11
pipreqs==0.4.13
psycopg==3.2.12
psycopg-binary==3.2.12
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
yarg==0.1.10
